Reduce some of the allocations
