+++ src/util/mod.rs
@@ -5,8 +5,6 @@ pub mod character_reference;
 pub mod constant;
 pub mod edit_map;
 pub mod encode;
-pub mod format_code_as_indented;
-pub mod format_heading_as_setext;
 pub mod gfm_tagfilter;
 pub mod identifier;
 pub mod infer;
